Remove option Update
Added the function to be able to remove an option from the product. Had
to change the "disabled" status of the all ready in place input fields
to "readonly" as this still allowed the value of the field to be passed
to the function.

Adjusted the database for the product_variation table to include an
"archive" column so that the options deleted would only be hidden but
still be able to viewed on orders.

Started work on showing the categories in the db on the categories page.